fix(RX/P7): 4 compile errors + CRITICAL defaultConf rtm_stop_distance_m
Compile fixes (from Claude.ai session):
1. PWM.ino: add (uint8_t) cast on rtm_steer_override to resolve ternary
   type mismatch (std::atomic<uint8_t> vs volatile uint8_t)
2. RTMState.ino: add extern bool gps_phase_b_ok; (defined in Radio.ino)
3. BREmote_V2_Rx.h: add rtm_stop_distance_m field to confStruct (fills
   former 2-byte tail padding; sizeof stays 152; static_assert updated)
4. ConfigService.ino: add rtm_stop_distance_m to kCfgFields[]

CRITICAL safety fix (found in security review 2026-04-26):
- defaultConf.rtm_stop_distance_m was missing from the initializer list.
  C++ zero-init would set it to 0, making Gate 9 check (dist_m < 0.0f)
  never fire — permanently disabling the hard stop that prevents the
  buggy from driving into the user. Fixed: add 3 (metres) to defaultConf.

Additional:
- BREmote_V2_Rx.h layout comment updated: tail padding line replaced with
  rtm_stop_distance_m entry
- RTMState.ino extern comment corrected: variable defined in Radio.ino,
  not GPS.ino

+++ Source/V2_Integration_Rx/PWM.ino
@@ -40,7 +40,7 @@ void calcPWM()
   // RTM can only subtract from user throttle (never add). Creator safety philosophy enforced.
   uint8_t effective_thr   = rtm_rx_emergency_stop ? 0 : thr_received;
   uint8_t effective_steer = (rtm_rx_active && usrConf.rtm_rx_override_steering)
-                            ? rtm_steer_override
+                            ? (uint8_t)rtm_steer_override
                             : steering_received;
 
   if(usrConf.steering_type == 0)
